+++ src/main/java/com/cipasoft/edule/repositories/RepoHour.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cipasoft.edule.models.Hour;
 
 @Repository
-public interface RepoHour extends JpaRepository<Hour, Long> {
+public interface RepoHour extends JpaRepository<Hour, Integer> {
     
 }
