+++ src/main/java/com/cipasoft/edule/services/ServUser.java
@@ -18,7 +18,7 @@ public class ServUser {
         return userRepository.findAll();
     }
 
-    public Optional<User> getUserById(Long id) {
+    public Optional<User> getUserById(Integer id) {
         return userRepository.findById(id);
     }
 
@@ -30,7 +30,7 @@ public class ServUser {
         return userRepository.save(user);
     }
 
-    public void deleteUser(Long id) {
+    public void deleteUser(Integer id) {
         userRepository.deleteById(id);
     }
 }
